% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdhg.R
\name{adaptive_pdhg}
\alias{adaptive_pdhg}
\title{Adaptive PDHG algorithm for group GMC}
\usage{
adaptive_pdhg(
  x,
  v,
  y,
  A,
  Z,
  group,
  K,
  lambda,
  maxiters = 1000,
  tol = 0.001,
  adaptive = TRUE,
  backtrack = TRUE,
  L,
  tau,
  sigma,
  a = 0.5,
  eta = 0.95,
  Delta = 2,
  gamma = 0.75,
  b = 0.95,
  stopRule = "ratioResidual",
  ShowTime = FALSE
)
}
\arguments{
\item{x}{The initial value for variable x, if missing, set a as zero vector}

\item{v}{The initial value for variable v, if missing, set as zero vector}

\item{y}{The response variable}

\item{A}{The A matrix in PDHG}

\item{Z}{The Z matrix in PDHG}

\item{group}{A vector describing the grouping of the coefficients.}

\item{K}{The weight vector for the group penalty or called group multiplier}

\item{lambda}{The tunning parameter in PDHG}

\item{maxiters}{The max number of iteration}

\item{tol}{The relative decrease in the residuals before the method stops}

\item{adaptive}{If 'TRUE' then use adaptive method.
If 'FALSE', turn off the adaptivity and users must set tau and
sigma as the setpsize for the classical pdhg}

\item{backtrack}{If 'TRUE' then use backtracking method.}

\item{L}{The reciprical spectral radius of Z'Z. If missing, the algorithm will give an approcimate value}

\item{tau}{The intial stepsize for the primal variables.
If missing, the algorithm will initialize it as the squared root of L}

\item{sigma}{The intial stepsize for the dual variables.
If missing, the algorithm will initialize it as L/tau}

\item{a}{Intial adaptive update strength for stepsizes}

\item{eta}{How fast does the adaptivity level decay}

\item{Delta}{Update stepsizes when primal/dual ratio exceeds Delta}

\item{gamma}{Used to determine when need to backtrack to maintain positivity conditions}

\item{b}{Adaptivity parameter used for backtracking update}

\item{stopRule}{Three choices in total. Default is ratioResidual
"ratioResidual": primal/maxPrimal<tol & dual/maxDual < tol
"residual": primal < tol & dual < tol
"iteration" iter > maxiters}

\item{ShowTime}{Whether to show the time cost for computing the solution}
}
\description{
Adaptive PDHG algorithm for group GMC
}
