% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_grGMC.R
\name{cv.grGMC}
\alias{cv.grGMC}
\title{Cross-validation for grouped variable selection with GMCpenalization}
\usage{
cv.grGMC(
  y,
  A,
  theta,
  group = 1:ncol(A),
  lambda,
  group.multiplier,
  Z.method = "theta",
  Z.epsilon = 0.001,
  nfolds = 10,
  seed = 1234,
  maxiters = 10000,
  tol = 1e-05,
  returnX = TRUE,
  returnY = FALSE,
  trace = FALSE
)
}
\arguments{
\item{y}{The response variable}

\item{A}{The design matrix without an intercept. grGMC standardizes the data and includes an intercept by default.}

\item{theta}{The convexity-preserving parameter theta in the group GMC penalty}

\item{group}{A vector describing the grouping of the coefficients}

\item{lambda}{The tunning parameter in grGMC, usually a user supplied sequence of decrasing values.}

\item{group.multiplier}{The weight vector for the group GMC penalty}

\item{Z.method}{Which method to set a matrix parameter Z = lambda B^TB/n}

\item{Z.epsilon}{The epsilon parameter in the new method of geenrating Z}

\item{nfolds}{The number of cross-validation folds. Default is 10.}

\item{seed}{Users can set the seed of the random number generator to obtain reproducible results.}

\item{maxiters}{The maxiters parameter in grGMC}

\item{tol}{The tol parameter in grGMC}

\item{returnX}{Whether to return the standardized data}

\item{returnY}{Whether to return the fitted values from the cross-validation folds}

\item{trace}{Whether to inform the users the progress of the cross-validation by announcing the beginning of each CV fold}
}
\value{
\code{cve} The error for each value of lambda, averaged across the cross-validation folds.

\code{cvse} The estimated standard error associated with each value of for cve.

\code{lambda} The sequence of regularization parameter values along which the cross-validation error was calculated.

\code{fit} The fitted grpreg object for the whole data.

\code{fold} The fold assignments for cross-validation for each observation.

\code{min} The index of lambda corresponding to lambda.min.

\code{min_1se} The index of lambda corresponding to lambda.1se.

\code{lambda.min} The value of lambda with respect to the minimum cross-validation error.

\code{lambda.1se} The value of lambda selected by the "one-standard-error" rule.
}
\description{
\code{cv.grGMC} performs k-fold cross-validation for penalizaed linear regression models with the groupe GMC penalty
over a grid of values for the regularization parameter lambda.
}
\examples{
set.seed(1234)
## set the ture signal
x_star=c(0, 3, 9, 0, 0, 15, 0, 4, 5, 0)
p=length(x_star)
##group the variables
group=c(2, 1, 1, 3, 2, 1, 3, 4, 4, 3)
##set sample size
n=15
## generate the covariate matrix
A=matrix(rnorm(n*p, mean = 1, sd=1), nrow = n)
## set the group weights
gp=unique(group)
J=length(gp)
K=rep(0,J)
for (j in 1:J) {
K[j]=sqrt(length(which(group==gp[j])))
}
## response vector
y=as.vector(A\%*\%x_star+rnorm(n, mean=0, sd = 0.5))
lambda= seq(1e-3, 1, length.out=20)
## fit with cv.grGMC function
fit_GMC <- cv.grGMC(y, A, theta=0.9, group, lambda, K)
}
\author{
Xiaoqian Liu
}
