% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.grGMC_socp.R
\name{cv.grGMC_socp}
\alias{cv.grGMC_socp}
\title{Cross-validation for grouped variable selection with GMC penalization}
\usage{
cv.grGMC_socp(
  y,
  A,
  theta,
  group = 1:ncol(A),
  lambda,
  group.multiplier,
  nfolds = 10,
  seed = 1234,
  maxiters = 10000,
  tol = 1e-05,
  returnX = TRUE,
  returnY = FALSE,
  trace = FALSE
)
}
\arguments{
\item{y}{The response variable}

\item{A}{The design matrix without an intercept. grGMC standardizes the data and includes an intercept by default.}

\item{theta}{The convexity-preserving parameter theta in the group GMC penalty}

\item{group}{A vector describing the grouping of the coefficients}

\item{lambda}{The tunning parameter in grGMC, usually a user supplied sequence of decrasing values.}

\item{group.multiplier}{The weight vector for the group GMC penalty}

\item{nfolds}{The number of cross-validation folds. Default is 10.}

\item{seed}{Users can set the seed of the random number generator to obtain reproducible results.}

\item{maxiters}{The maxiters parameter in grGMC}

\item{tol}{The tol parameter in grGMC}

\item{returnX}{Whether to return the standardized data}

\item{returnY}{Whether to return the fitted values from the cross-validation folds}

\item{trace}{Whether to inform the users the progress of the cross-validation by announcing the beginning of each CV fold}
}
\value{
\code{cve} The error for each value of lambda, averaged across the cross-validation folds.

\code{cvse} The estimated standard error associated with each value of for cve.

\code{lambda} The sequence of regularization parameter values along which the cross-validation error was calculated.

\code{fit} The fitted grpreg object for the whole data.

\code{fold} The fold assignments for cross-validation for each observation.

\code{min} The index of lambda corresponding to lambda.min.

\code{min_1se} The index of lambda corresponding to lambda.1se.

\code{lambda.min} The value of lambda with respect to the minimum cross-validation error.

\code{lambda.1se} The value of lambda selected by the "one-standard-error" rule.
}
\description{
\code{cv.grGMC_socp} performs k-fold cross-validation for penalizaed linear regression models with the groupe GMC penalty
over a grid of values for the regularization parameter lambda using the SOCP solver.
}
\author{
Xiaoqian Liu
}
