% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_grGMC}
\alias{df_grGMC}
\title{DOF function in "Penalized methods for bi-level variable selection"
 Patrick Breheny and Jian Huang, 2009
 Modified version}
\usage{
df_grGMC(y, A, x, group, df.method = "group")
}
\arguments{
\item{y}{response, "centered"}

\item{A}{covariate matrix, "standardized"}

\item{x}{the current estimate, intercept is not included}

\item{group}{A vector describing the grouping of the coefficients}

\item{df.method}{which method to compute the degrees of freedom,
"group" means df=number of nonzero groups,
"active" means df=number of nonzero coefficients,
"Breheny" means using the formula in Breheny(2009)}
}
\description{
DOF function in "Penalized methods for bi-level variable selection"
 Patrick Breheny and Jian Huang, 2009
 Modified version
}
