% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinOffDiag.R
\name{MinOffDiag}
\alias{MinOffDiag}
\title{\code{MinOffDiag} The new method to find a feasible matrix D by controlling the off-diagonal elements}
\usage{
MinOffDiag(D0, U10, U20, U30, M, epsilon, maxiters = 1000, tol = 0.001)
}
\arguments{
\item{D0}{the initial matrix we want to deal with}

\item{U10}{the initial value}

\item{U20}{the initial value}

\item{U30}{the initial value}

\item{M}{the fixed M matrix}

\item{epsilon}{the largest value of the off-diagonal elements}

\item{maxiters}{the maximal number of iteration}

\item{tol}{the tolerance to terminate the algorithm}
}
\value{
\code{D} the feasible matrix D

\code{iters} the number of iteration used

\code{diff} the difference when terminate the algorithm
}
\description{
\code{MinOffDiag} The new method to find a feasible matrix D by controlling the off-diagonal elements
}
