% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_socp.R
\name{obj_socp}
\alias{obj_socp}
\title{\code{obj_socp} computes the objective value of the group GMC problem as a SOCP}
\usage{
obj_socp(y, A, Z, beta, lambda, t, group, K)
}
\arguments{
\item{y}{The response variable}

\item{A}{The design matrix without an intercept.}

\item{Z}{The Z matrix in the group GMC problem}

\item{beta}{The specific solution of beta}

\item{lambda}{The tunning parameter in grGMC}

\item{t}{The specific solution of t}

\item{group}{A vector describing the grouping of the coefficients}

\item{K}{The weight vector for the group GMC penalty = group.multiplier}
}
\description{
\code{obj_socp} computes the objective value of the group GMC problem as a SOCP
}
\author{
Xiaoqian Liu
}
