% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_grGMC.R
\name{objective_grGMC}
\alias{objective_grGMC}
\title{\code{objective_grGMC} computes the objective value of the group GMC problem at a specific solution point}
\usage{
objective_grGMC(y, A, Z, beta, v, lambda, group, group.multiplier)
}
\arguments{
\item{y}{The response variable}

\item{A}{The design matrix without an intercept.}

\item{Z}{The Z matrix in the group GMC problem}

\item{beta}{The specific solution of beta}

\item{v}{The specific solution of v}

\item{lambda}{The tunning parameter in grGMC}

\item{group}{A vector describing the grouping of the coefficients}

\item{group.multiplier}{The weight vector for the group GMC penalty}
}
\description{
\code{objective_grGMC} computes the objective value of the group GMC problem at a specific solution point
}
\author{
Xiaoqian Liu
}
