% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinOffDiag.R
\name{Proj_C2}
\alias{Proj_C2}
\title{\code{Proj_C2} Project a matrix D to the set M-D is positive semidefinite}
\usage{
Proj_C2(D, M, tol = 0)
}
\arguments{
\item{D}{the matrix we want to deal with}

\item{M}{the fixed M matrix}

\item{tol}{the minimum eigenvalue = 0}
}
\value{
\code{D_new} The projected matrix
}
\description{
\code{Proj_C2} Project a matrix D to the set M-D is positive semidefinite
}
