% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinOffDiag.R
\name{Proj_C3}
\alias{Proj_C3}
\title{\code{Proj_C3} Project a matrix D onto the set where M-D <= epsilon for off-diagonal elements}
\usage{
Proj_C3(D, M, epsilon)
}
\arguments{
\item{D}{the matrix we want to deal with}

\item{M}{the fixed M matrix}

\item{epsilon}{the largest value of the off-diagonal elements}
}
\value{
\code{D_new} The projected matrix
}
\description{
\code{Proj_C3} Project a matrix D onto the set where M-D <= epsilon for off-diagonal elements
}
