% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinOffDiag.R
\name{Proj_PSM}
\alias{Proj_PSM}
\title{\code{Proj_PSM} Project a matrix to the positive semidefinite matrix set}
\usage{
Proj_PSM(D, tol = 0)
}
\arguments{
\item{D}{the matrix we want to deal with}

\item{tol}{the minimum eigenvalue = 0}
}
\value{
\code{D_new} The projected matrix
}
\description{
\code{Proj_PSM} Project a matrix to the positive semidefinite matrix set
}
