% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socp_solver.R
\name{socp_solver}
\alias{socp_solver}
\title{The SOCP solver for group GMC}
\usage{
socp_solver(X, y, Z, theta, lambda, group, K, maxiters = 10000, tol = 1e-12)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response variable}

\item{Z}{The Z matrix in the group GMC problem}

\item{theta}{The convexity-preserving parameter theta in the group GMC penalty}

\item{lambda}{The tunning parameter in grGMC, usually a user supplied sequence of decrasing values.}

\item{group}{A vector describing the grouping of the coefficients}

\item{K}{The weight vector for the group GMC penalty = group.multiplier}

\item{maxiters}{The max number of iterations for the SOCP solver}

\item{tol}{The tolerance to terminate the SOCP solver}
}
\description{
The SOCP solver for group GMC
}
\author{
Xiaoqian Liu
}
